refactor(dccpeep): extract pass_once dispatcher into peep_pass_once.c
Move the single-scan micro-pattern dispatcher pass_once() and its 21 private try_*_at helpers (plus local_alloc_hl_result_dead, is_jp_to_next_label, replace_block_with_5, try_fold_bool_branch, and peep_line_in_function, which are now used only there) into peep_pass_once.c. Promote the genuinely shared predicates is_uncond_jp, label_name_at, line_is_label_name, and peep_is_public_line to the internal header.

dccpeep.c drops from ~12.3k to ~10.6k lines. Validation: strict build clean; all 316 raw inputs byte-identical in default, -Os, and -fundocumented-z80 modes; direct fixtures pass.
